% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Germany.1918}
\alias{Germany.1918}
\title{Germany 1918 dataset}
\format{
A vector of numeric values named with date of record.
}
\usage{
data(Germany.1918)
}
\description{
This dataset is used as an example in the \code{R0} package to estimate reproduction
ratios using the available methods. It is the emporal distribution of Spanish
flu in Prussia, Germany, from the 1918-19 influenza epidemic.
}
\references{
Peiper O: Die Grippe-Epidemie in Preussen im Jahre 1918/19. Veroeffentlichungen aus dem Gebiete der Medizinalverwaltung 1920, 10:  417-479 (in German). \cr
Nishiura H. Time variations in the transmissibility of pandemic influenza in Prussia, Germany, from 1918-19. In: Theoretical Biology and Medical Modelling
}
\keyword{datasets}
