% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{get.scale}
\alias{get.scale}
\title{Internal scaling function to display proper X-Axis labels}
\usage{
get.scale(scale)
}
\description{
Internal scaling function to display proper X-Axis labels
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
