\name{plot.R0.GT}
\alias{plot.R0.GT}
\title{Print the characteristics of the generation time distribution}
\description{Prints the characteristics of the generation time distribution}
\usage{\method{plot}{R0.GT}(x, ...)}
\arguments{
  \item{x}{the generation time distribution.}
  \item{\dots}{extra parameters passed to plot.}
}
\details{For internal use. Called by print.}
\value{Called for side effect. Shows a plot of the generation time distribution.	}

\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
