% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.R.R
\name{plotfitRSB}
\alias{plotfitRSB}
\title{Internal plotfit method for AR estimates}
\usage{
plotfitRSB(x, xscale, SB.dist, ...)
}
\arguments{
\item{x}{An output of \code{\link[=est.R0.SB]{est.R0.SB()}} (class \code{R0.R}).}

\item{xscale}{Scale to be adjusted on x-axis. Can be \code{d} (day), \code{w} (week (default)), \code{f} (fornight), \code{m} (month).}

\item{SB.dist}{Boolean. Should the R distirbution throughout the epidemic be plotted for the SB method (defaults to \code{TRUE}) ?}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect. Draws the fit of one estimation method to the data.
}
\description{
Plot the fit of a single model output to epidemic data when the method is SB.
}
\details{
For internal use. This function is called by the \code{\link[=plotfit.R0.R]{plotfit.R0.R()}}.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
