% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.R.R
\name{plotfitRxx}
\alias{plotfitRxx}
\title{Internal plotfit method for EG, ML and TD estimates}
\usage{
plotfitRxx(x, xscale, ...)
}
\arguments{
\item{x}{An output of \code{\link[=est.R0.EG]{est.R0.EG()}}, \code{\link[=est.R0.ML]{est.R0.ML()}} or \code{\link[=est.R0.TD]{est.R0.TD()}} (class \code{R0.R}).}

\item{xscale}{Scale to be adjusted on x-axis. Can be \code{d} (day), \code{w} (week (default)), \code{f} (fornight), \code{m} (month).}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect. Draws the fit of one estimation method to the data.
}
\description{
Plot the fit of a single model output to epidemic data when the method is
EG, ML or TD.
}
\details{
For internal use. This function is called by the \code{\link[=plotfit.R0.R]{plotfit.R0.R()}}.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
