% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRAR}
\alias{plotRAR}
\title{Plot R0 for Attack Rate method}
\usage{
plotRAR(x, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{\link[=estimate.R]{estimate.R()}} with \code{method = "AR"}.}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect.
}
\description{
Internal plot method for AR estimates.
}
\details{
This plots the R0 reproduction number along with confidence interval.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
