% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotREG}
\alias{plotREG}
\title{Internal plot method for EG estimates}
\usage{
plotREG(x, ...)
}
\description{
Internal plot method for EG estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
