% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRML}
\alias{plotRML}
\title{Internal plot method for ML estimates}
\usage{
plotRML(x, ...)
}
\description{
Internal plot method for ML estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
