% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRSB}
\alias{plotRSB}
\title{Internal plot method for SB estimates}
\usage{
plotRSB(x, xscale, ...)
}
\description{
Internal plot method for SB estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
