% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRTD}
\alias{plotRTD}
\title{Internal plot method for TD estimates}
\usage{
plotRTD(x, xscale, TD.split, ...)
}
\description{
Internal plot method for TD estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
