% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R0.GT.R
\name{print.R0.GT}
\alias{print.R0.GT}
\title{Print the characteristics of the generation time distribution}
\usage{
\method{print}{R0.GT}(x, ...)
}
\arguments{
\item{x}{the GT distribution.}

\item{\dots}{Parameters passed to inner functions}
}
\value{
Called for side effect. Displays GT and mean/sd.
}
\description{
Prints the characteristics of the generation time distribution.
}
\details{
For internal use. Called by print.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
