% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R0.R.R
\name{print.R0.R}
\alias{print.R0.R}
\title{Print the R estimate}
\usage{
\method{print}{R0.R}(x, ...)
}
\arguments{
\item{x}{Result of est.R0.xx (class R0.R)}

\item{\dots}{Parameters passed to inner functions}
}
\value{
Called for side effect.
}
\description{
Prints the R estimate
}
\details{
For internal use. Called by print.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
