\name{R0-package}
\alias{R0-package}
\alias{R0}
\docType{package}
\title{Estimation of R0 and real-time reproduction number from epidemics}
\description{This package is designed to estimate reproduction numbers for disease outbreak, based on incidence data. It implements several documented methods. It is therefore possible to compare estimations according to the methods used. Depending on the methods requested by user, basic reproduction number (commonly denoted as R0) or real-time reproduction number (referred to as R(t)) is computed, along with a 95\% Confidence Interval. Plotting outputs will give different graphs depending on the methods requested : basic reproductive number estimations will only show the epidemic curve (collected data) and an adjusted model, whereas real-time methods will also show the R(t) variations throughout the outbreak time period. Sensitivity analysis tools are also provided, and allow for investigating effects of varying Generation Time distribution or time window on estimates.}
\details{
\tabular{ll}{Package: \tab R0\cr
Type: \tab Package\cr
Title: \tab Estimation of R0 and real-time reproduction number from epidemics\cr
Version: \tab 1.2-1\cr
Date: \tab 2013-07-08\cr
Author: \tab Pierre-Yves Boelle, Thomas Obadia\cr
Maintainer: \tab Thomas Obadia <obadia@u707.jussieu.fr>\cr
Depends: \tab R (>= 2.13.0), MASS\cr
Suggests: \tab epicalc\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr}
}
\author{Pierre-Yves Boelle, Thomas Obadia}

\keyword{ package }


