% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Germany.1918}
\alias{Germany.1918}
\title{Germany.1918 exemple dataset}
\format{
The format is: num [1:126] 10 4 4 19 6 13 28 23 35 27 ...
}
\source{
Peiper O: Die Grippe-Epidemie in Preussen im Jahre 1918/19.
Veroeffentlichungen aus dem Gebiete der Medizinalverwaltung 1920, 10:
417-479 (in German).
}
\description{
Temporal distribution of Spanish flu in Prussia, Germany, from 1918-19
}
\examples{

data(Germany.1918)
## maybe str(Germany.1918) ; plot(Germany.1918) ...

}
\references{
Nishiura H. Time variations in the transmissibility of pandemic
influenza in Prussia, Germany, from 1918-19. In: Theoretical Biology and
Medical Modelling
}
\keyword{datasets}
