\name{Germany.1918}
\alias{Germany.1918}
\docType{data}
\title{
Germany.1918 exemple dataset
}
\description{
Temporal distribution of Spanish flu in Prussia, Germany, from 1918-19
}
\usage{data(Germany.1918)}
\format{
  The format is:
 num [1:126] 10 4 4 19 6 13 28 23 35 27 ...
}
\source{
Peiper O: Die Grippe-Epidemie in Preussen im Jahre 1918/19. Veroeffentlichungen aus dem Gebiete der Medizinalverwaltung 1920, 10:  417-479 (in German).
}
\references{
Nishiura H. Time variations in the transmissibility of pandemic influenza in Prussia, Germany, from 1918-19. In: Theoretical Biology and Medical Modelling
}
\examples{
data(Germany.1918)
## maybe str(Germany.1918) ; plot(Germany.1918) ...
}
\keyword{datasets}
