\name{get.scale}
\alias{get.scale}
\title{Internal scaling function to display proper X-Axis labels}
\description{Internal scaling function to display proper X-Axis labels}
\usage{get.scale(scale)}
\arguments{
  \item{scale}{
}
}



\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
