% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{get.scale}
\alias{get.scale}
\title{Scaling of x-axis}
\usage{
get.scale(scale)
}
\arguments{
\item{scale}{Scale to be adjusted on x-axis. Can be \code{d} (day), \code{w} (week (default)), \code{f} (fornight), \code{m} (month).}
}
\value{
An integer corresponding to the number of days between each x-axis tickmark.
}
\description{
Internal scaling function to display proper x-axis labels.
}
\details{
Builds the x-axis labels corresponding to a human-friendly level (day, week...).
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
