\name{GT.chld.hsld}
\alias{GT.chld.hsld}
\docType{data}
\title{
2009 A/H1N1 observed Generation Time distribution
}
\description{
Observed generation time distribution for children in household for the 2009 A/H1N1 influenza pandemic.
}
\usage{data(GT.chld.hsld)}
\format{
  The format is:
 num [1:8] 0 0.25 0.2 0.15 0.1 0.09 0.05 0.01
}
\source{
S. Cauchemez, A. Bhattarai, T.L. Marchbanks, R.P. Fagan, S. Ostroff, N.M. Ferguson, et al., Role of Social Networks in Shaping Disease Transmission During a Community Outbreak of 2009 H1N1 Pandemic Influenza, Pnas. 108 (2011) 2825-2830.
}
\examples{
data(GT.chld.hsld)
## maybe str(GT.chld.hsld) ; plot(GT.chld.hsld) ...
}
\keyword{datasets}
