% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GT.chld.hsld}
\alias{GT.chld.hsld}
\title{2009 A/H1N1 observed generation time distribution}
\format{
A vector of unnamed numeric values.
}
\usage{
data(GT.chld.hsld)
}
\description{
Observed generation time distribution for children in household for the 2009
A/H1N1 influenza pandemic.
}
\references{
S. Cauchemez, A. Bhattarai, T.L. Marchbanks, R.P. Fagan, S. Ostroff, N.M. Ferguson, et al., Role of Social Networks in Shaping Disease Transmission During a Community Outbreak of 2009 H1N1 Pandemic Influenza, Pnas. 108 (2011) 2825-2830.
}
\keyword{datasets}
