% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.GT.R
\name{plot.R0.GT}
\alias{plot.R0.GT}
\title{Print the characteristics of the generation time distribution}
\usage{
\method{plot}{R0.GT}(x, ...)
}
\arguments{
\item{x}{the generation time distribution.}

\item{\dots}{extra parameters passed to plot.}
}
\value{
Called for side effect. Shows a plot of the generation time
distribution.
}
\description{
Prints the characteristics of the generation time distribution
}
\details{
For internal use. Called by print.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
