% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.GT.R
\name{plot.R0.GT}
\alias{plot.R0.GT}
\title{Plot a generation time distribution}
\usage{
\method{plot}{R0.GT}(x, ...)
}
\arguments{
\item{x}{Generation time distribution from \code{\link[=generation.time]{generation.time()}}.}

\item{...}{Parameters passed to inner functions.}
}
\value{
Called for side effect. Shows a plot of the generation time distribution.
}
\description{
Plots the distribution of a generation time object from \code{\link[=generation.time]{generation.time()}}.
}
\details{
For internal use. Called by \code{\link[base:print]{base::print()}}.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
