% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.sR.R
\name{plot.R0.sR}
\alias{plot.R0.sR}
\title{Plot the R0/Rt value along with confidence interval of all requested models
to epidemic data}
\usage{
\method{plot}{R0.sR}(x, xscale = "w", TD.split = FALSE, ...)
}
\arguments{
\item{x}{Result of est.R0 (class sR)}

\item{xscale}{Scale to be adjusted on X axis. Can be "d" (day), "w" (week
(default)), "f" (fornight), "m" (month).}

\item{TD.split}{Parameter to force the display of both R(t) and the epidemic
curve in the same window for TD method}

\item{\dots}{parameters passed to plot.R}
}
\value{
Called for its side effect : Draws all R0 or R(t) values from
requested estimation methods.
}
\description{
Plot the R0/Rt value along with confidence interval of all requested models
to epidemic data
}
\details{
For internal use. Called by plot.

Tweaked plot() function that draws the reproduction number values for each
method contained in the object constructed by est.RO().
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
