% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.sR.R
\name{plotfit.R0.sR}
\alias{plotfit.R0.sR}
\title{Plot a model fit for \code{R0.sR} objects}
\usage{
\method{plotfit}{R0.sR}(x, all = TRUE, xscale = "w", SB.dist = TRUE, ...)
}
\arguments{
\item{x}{An output of \code{\link[=estimate.R]{estimate.R()}} (class \code{R0.sR}).}

\item{all}{Boolean. Should the full epidemic curve be shown ?}

\item{xscale}{Scale to be adjusted on x-axis. Can be \code{d} (day), \code{w} (week (default)), \code{f} (fornight), \code{m} (month).}

\item{SB.dist}{Boolean. Should the R distirbution throughout the epidemic be plotted for the SB method (defaults to \code{TRUE}) ?}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect. Draws the fit of one or more estimation method to the
data.
}
\description{
Plot the fit of model outputs to epidemic data.
}
\details{
For internal use. This function is called by the \code{\link[=plotfit]{plotfit()}} S3 method.
It will sequentialy call the necesary \code{\link[=plotfit]{plotfit()}} methods for each estimation
contained in the \code{x} argument.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
