% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.sR.R
\name{plotfit.R0.sR}
\alias{plotfit.R0.sR}
\title{Plot the fit of all requested models to epidemic data}
\usage{
\method{plotfit}{R0.sR}(x, all = TRUE, xscale = "w", SB.dist = TRUE, ...)
}
\arguments{
\item{x}{Result of est.R (class R0.R)}

\item{all}{Should the whole epidemic curve be shown}

\item{xscale}{Scale to be adjusted on X axis. Can be "d" (day), "w" (week
(default)), "f" (fornight), "m" (month).}

\item{SB.dist}{Should R distribution throughout the epidemic be plotted for
SB method? (default: TRUE)}

\item{\dots}{Parameters passed to plot}
}
\value{
Called for its side effect : Draws \code{all} the epidemic curves
and associated fit data computed by est.R0
}
\description{
Plots the fit of \code{all} requested models to epidemic data
}
\details{
For internal use. Called by \code{\link{plotfit}}.

Tweaked plot() function that draws the epidemic data and model fit of each
method contained in the object constructed by est.RO().
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
