\name{plotfitRAR}
\alias{plotfitRAR}
\title{Internal plotfit method for AR estimates}
\description{Internal plotfit method for AR estimates}
\usage{plotfitRAR(x, xscale, ...)}
\arguments{
  \item{x}{
}
  \item{xscale}{
}
  \item{\dots}{
}
}



\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
