% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.R.R
\name{plotfitRAR}
\alias{plotfitRAR}
\title{Internal plotfit method for AR estimates}
\usage{
plotfitRAR(x, xscale, ...)
}
\description{
Internal \code{\link{plotfit}} method for AR estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
