% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.R.R
\name{plotfitRSB}
\alias{plotfitRSB}
\title{Internal plot method for SB estimates}
\usage{
plotfitRSB(x, xscale, SB.dist, ...)
}
\description{
Internal \code{\link{plotfit}} method for SB estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
