\name{plotfitRSB}
\alias{plotfitRSB}
\title{Internal plot method for SB estimates}
\description{Internal \code{\link{plotfit}} method for SB estimates}
\usage{plotfitRSB(x, xscale, 
    SB.dist, ...)}
\arguments{
  \item{x}{
}
  \item{xscale}{
}
  \item{SB.dist}{
}
  \item{\dots}{
}
}



\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
