% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfit.R0.R.R
\name{plotfitRxx}
\alias{plotfitRxx}
\title{Internal plotfit method for EG, ML and TD estimates}
\usage{
plotfitRxx(x, xscale, ...)
}
\description{
Internal \code{\link{plotfit}} method for EG, ML and TD estimates
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
