\name{plotfitRxx}
\alias{plotfitRxx}
\title{Internal plotfit method for EG, ML and TD estimates}
\description{Internal plotfit method for EG, ML and TD estimates}
\usage{plotfitRxx(x, xscale, ...)}
\arguments{
  \item{x}{
}
  \item{xscale}{
}
  \item{\dots}{
}
}



\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
