% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotREG}
\alias{plotREG}
\title{Plot R0 for Exponential Growth method}
\usage{
plotREG(x, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{\link[=estimate.R]{estimate.R()}} with \code{method = "EG"}.}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect.
}
\description{
Internal plot method for EG estimates.
}
\details{
This plots the exponential growth parameter r, along with R and
corresponding confidence interval.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
