% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRML}
\alias{plotRML}
\title{Plot R0 for Maximum Likelihood method}
\usage{
plotRML(x, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{\link[=estimate.R0]{estimate.R0()}} with \code{method = "ML"}.}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect.
}
\description{
Internal plot method for ML estimates.
}
\details{
This plots the R0 reproduction number along with confidence interval.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
