% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRSB}
\alias{plotRSB}
\title{Plot R0 for Sequential Bayesian method}
\usage{
plotRSB(x, xscale, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{\link[=estimate.R0]{estimate.R0()}} with \code{method = "SB"}.}

\item{xscale}{Scale to be adjusted on x-axis. Can be \code{d} (day), \code{w} (week (default)), \code{f} (fornight), \code{m} (month).}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect.
}
\description{
Internal plot method for SB estimates.
}
\details{
This plots the bayesian estimation of \eqn{R_{0}} along with its
credible interval.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
