% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R0.R.R
\name{plotRTD}
\alias{plotRTD}
\title{Plot R0 for Time-Dependent method}
\usage{
plotRTD(x, xscale, TD.split, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{\link[=estimate.R]{estimate.R()}} with \code{method = "TD"}.}

\item{xscale}{Scale to be adjusted on x-axis. Can be \code{d} (day), \code{w} (week (default)), \code{f} (fornight), \code{m} (month).}

\item{TD.split}{Boolean. Parameter to force the display of both \eqn{R(t)} and the epidemic curve in the same window for the TD method.}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effect.
}
\description{
Internal plot method for TD estimates.
}
\details{
This plots the time-dependent \eqn{R(t)} reproduction number along with
confidence band.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
