\name{plotRTD}
\alias{plotRTD}
\title{Internal plot method for TD estimates}
\description{Internal plot method for TD estimates}
\usage{plotRTD(x, xscale, TD.split, ...)}
\arguments{
  \item{x}{
}
  \item{xscale}{
}
  \item{TD.split}{
}
  \item{\dots}{
}
}



\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
