\name{print.R0.GT}
\alias{print.R0.GT}
\title{Print the characteristics of the generation time distribution}
\description{Prints the characteristics of the generation time distribution.}
\usage{\method{print}{R0.GT}(x, ...)}
\arguments{
  \item{x}{the GT distribution.}
  \item{\dots}{Parameters passed to inner functions}
}
\details{For internal use. Called by print.}
\value{Called for side effect. Displays GT and mean/sd.}

\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
