% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R0.R.R
\name{print.R0.R}
\alias{print.R0.R}
\title{Print method for objects of class \code{R0.R}}
\usage{
\method{print}{R0.R}(x, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{est.R0.xx()} routines (class \code{R0.R}).}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effects.
}
\description{
Prints summary statistics for an estimated reproduction ratio.
}
\details{
For internal use.

Displays the estimated reproduction ratio along with its confidence interval.
For the TD method, the time-series of \eqn{R(t)} is printed.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
