\name{print.R0.R}
\alias{print.R0.R}
\title{Print the R estimate}
\description{Prints the R estimate}
\usage{\method{print}{R0.R}(x, ...)}
\arguments{
  \item{x}{Result of est.R0.xx (class R0.R)}
  \item{\dots}{Parameters passed to inner functions}
}
\details{For internal use. Called by print.}
\value{Called for side effect.}

\author{Pierre-Yves Boelle, Thomas Obadia}






\keyword{internal}
