% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R0.sR.R
\name{print.R0.sR}
\alias{print.R0.sR}
\title{Print method for objects of class \code{R0.sR}}
\usage{
\method{print}{R0.sR}(x, ...)
}
\arguments{
\item{x}{An estimated object, output from \code{\link[=estimate.R]{estimate.R()}} (class \code{R0.sR}).}

\item{...}{Parameters passed to inner functions.}
}
\value{
This function does not return any data.
Called for side effects.
}
\description{
Prints summary statistics from \code{\link[=estimate.R]{estimate.R()}} outputs.
}
\details{
For internal use.

Displays the estimated reproduction ratio of one or more estimation results
generated by \code{\link[=estimate.R]{estimate.R()}}.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
