% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.from.r.R
\name{R.from.r}
\alias{R.from.r}
\title{Compute the discretized Laplace Transform using a discretized Generation
Time distribution}
\usage{
R.from.r(r, GT)
}
\arguments{
\item{r}{exponential growth ratio}

\item{GT}{discretized generation time distribution}
}
\value{
An R value corresponding to inverse of discretized Laplace
transform.
}
\description{
Computes the discretized Laplace Transform using a discretized Generation
Time distribution.
}
\details{
For internal use. Called by \code{\link{est.R0.EG}}.
}
\note{
The formula for the discretized Laplace transform is taken from
Wallinga.
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{internal}
