\name{R0-package}
\alias{R0-package}
\alias{R0}
\docType{package}
\title{
\packageTitle{R0}
}
\description{
\packageDescription{R0}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{R0}
\packageIndices{R0}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{R0}

Maintainer: \packageMaintainer{R0}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
