% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{R0-package}
\alias{R0-package}
\alias{R0}
\title{Estimation of R0 and Real-Time Reproduction Number from Epidemics}
\description{
Estimation of reproduction numbers for disease outbreak, based on incidence
data. The R0 package implements several documented methods. It is therefore
possible to compare estimations according to the methods used. Depending on
the methods requested by user, basic reproduction number (commonly denoted
as R0) or real-time reproduction number (referred to as R(t)) is computed,
along with a 95\% Confidence Interval. Plotting outputs will give different
graphs depending on the methods requested : basic reproductive number
estimations will only show the epidemic curve (collected data) and an
adjusted model, whereas real-time methods will also show the R(t) variations
throughout the outbreak time period. Sensitivity analysis tools are also
provided, and allow for investigating effects of varying Generation Time
distribution or time window on estimates.
}
\details{
\tabular{ll}{Package: \tab R0\cr Type: \tab Package\cr Title: \tab
Estimation of R0 and Real-Time Reproduction Number from Epidemics\cr
Version: \tab 1.2-6\cr Date: \tab 2015-05-21\cr Author: \tab Pierre-Yves
Boelle, Thomas Obadia\cr Maintainer: \tab Thomas Obadia
<thomas.obadia@iplesp.upmc.fr>\cr Depends: \tab R (>= 2.13.0), MASS\cr
License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr}
}
\author{
Pierre-Yves Boelle, Thomas Obadia
}
\keyword{package}
